// src/types/nfts.ts
/*-
 *
 * Hedera NFT SDK
 *
 * Copyright (C) 2024 Hedera Hashgraph, LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */
export interface NFTDetails {
  account_id: string;
  created_timestamp: string;
  delegating_spender?: null | string;
  deleted: boolean;
  metadata: string;
  modified_timestamp: string;
  serial_number: number;
  token_id: string;
  spender: null | string;
}

export interface NFTTransactions {
  consensus_timestamp: string;
  nonce: number;
  transaction_id: string;
  type: string;
  is_approval: boolean;
  receiver_account_id: string;
  sender_account_id: string;
}

export interface NFTTransactionsRequest {
  transactions: NFTTransactions[];
  links: Links;
}

interface Links {
  next: string | null;
}

export interface NFTS {
  nfts: NFTDetails[];
  links: Links;
}

export interface DecodedMetadata {
  metadata: string;
  serialNumber: number;
}

export interface ParsedNFTS {
  decodedMetadata: DecodedMetadata[];
  links: Links;
}